GEFP-19-AA kommenterat ut kopplingen till match. även OrphineRemoval på game. Skapat metod för att kapa kopplingarna till andra tabeller och uppdatera deras listor. Klar med removePlayer.

+++ src/main/java/com/example/piedpiperdb/DAO/PlayerDAO.java
@@ -1,6 +1,9 @@
 package com.example.piedpiperdb.DAO;
 
+import com.example.piedpiperdb.Entities.Game;
+import com.example.piedpiperdb.Entities.Match;
 import com.example.piedpiperdb.Entities.Player;
+import com.example.piedpiperdb.Entities.Team;
 import jakarta.persistence.*;
 
 import java.util.ArrayList;
@@ -132,16 +135,36 @@ public class PlayerDAO {
     }
 
     public boolean deletePlayerById(int id) {
+        if (!removePlayerRelations(id)){
+            System.out.println("Could not remove player relations. Player is not deleted.");
+            return false;
+        }
+
         EntityManager entityManager = ENTITY_MANAGER_FACTORY.createEntityManager();
         EntityTransaction transaction = null;
 
         try{
             transaction = entityManager.getTransaction();
             transaction.begin();
+
             Player playerToDelete = entityManager.find(Player.class, id);
+            if (playerToDelete == null) {
+                System.out.println("Player not found in database.");
+                transaction.rollback();
+                return false;
+            }
+
             entityManager.remove(entityManager.contains(playerToDelete) ? playerToDelete : entityManager.merge(playerToDelete));
             transaction.commit();
-            return true;
+            playerToDelete = entityManager.find(Player.class, id);
+            if(playerToDelete == null) {
+                System.out.println("Player deleted from DB");
+                return true;
+            } else {
+                System.out.println("Player still exist in DB" + playerToDelete.toString());
+                return false;
+            }
+
         } catch (Exception e) {
             System.out.println(e.getMessage());
             if (entityManager != null && transaction != null && transaction.isActive()) {
@@ -152,4 +175,60 @@ public class PlayerDAO {
             entityManager.close();
         }
     }
+
+    //GBFP-19-AA
+    public boolean removePlayerRelations(int playerId) {
+        EntityManager entityManager = ENTITY_MANAGER_FACTORY.createEntityManager();
+        EntityTransaction transaction = null;
+
+        try {
+            transaction = entityManager.getTransaction();
+            transaction.begin();
+
+            Player player = entityManager.find(Player.class, playerId);
+            if(player == null) {
+                System.out.println("Player not found in the DB" + playerId);
+                return false;
+            } else {
+                System.out.println("Removing relations...");
+                if (player.getGameId() != null) {
+                    Game game = player.getGameId();
+                    System.out.println("Removing player from game" + game);
+                    game.getPlayers().remove(player);
+                    player.setGameId(null);
+                    entityManager.merge(game);
+                }
+
+                if (player.getTeamId() != null) {
+                    Team team = player.getTeamId();
+                    System.out.println("Removing player from team" + team);
+                    team.getListOfPlayersInTeam().remove(player);
+                    player.setTeamId(null);
+                    entityManager.merge(team);
+                }
+
+/*                if (player.getMatchId() != null) {
+                    Match match = player.getMatchId();
+                    System.out.println("Removing player from match" + match);
+                    //match.getListOfPlayers().remove(player);
+                    player.setMatchId(null);
+                    entityManager.merge(match);
+                }*/
+
+                entityManager.merge(player);
+                transaction.commit();
+                System.out.println("Player relations removed from DB");
+                return true;
+            }
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
+            if (entityManager != null && transaction != null && transaction.isActive()) {
+                transaction.rollback();
+            }
+            return false;
+        } finally {
+            entityManager.close();
+        }
+    }
+
 }
